refactor: moved unit test to test folder and useCallbacks
