+++ frontend/src/components/ExplorerCard/test/ExplorerCard.test.tsx
@@ -2,8 +2,8 @@ import { fireEvent, render, screen } from '@testing-library/react';
 import ROUTES from 'constants/routes';
 import { QueryClient, QueryClientProvider } from 'react-query';
 
-import { viewMockData } from './__mock__/viewData';
-import ExplorerCard from './ExplorerCard';
+import { viewMockData } from '../__mock__/viewData';
+import ExplorerCard from '../ExplorerCard';
 
 jest.mock('react-router-dom', () => ({
 	...jest.requireActual('react-router-dom'),
