+++ frontend/src/components/ExplorerCard/test/MenuItemGenerator.test.tsx
@@ -2,8 +2,8 @@ import { fireEvent, render, screen } from '@testing-library/react';
 import ROUTES from 'constants/routes';
 import { QueryClient, QueryClientProvider } from 'react-query';
 
-import { viewMockData } from './__mock__/viewData';
-import MenuItemGenerator from './MenuItemGenerator';
+import { viewMockData } from '../__mock__/viewData';
+import MenuItemGenerator from '../MenuItemGenerator';
 
 jest.mock('react-router-dom', () => ({
 	...jest.requireActual('react-router-dom'),
